backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
pytz==2020.5
PyYAML==5.4.1
sentry-sdk==0.20.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
